requirements.txt
# Auto-generated: PyTorch installed separately via setup.sh
# Use: uv pip install -r requirements.txt -i https://pypi.org/simple/

transformers>=4.46.0
datasets>=2.21.0
accelerate>=0.34.0
numpy>=1.26.0
scipy>=1.14.0
matplotlib>=3.9.0
tqdm>=4.66.0
pandas>=2.2.0
pyyaml>=6.0.1
huggingface_hub>=0.25.0

# Training
trl>=0.12.0
peft>=0.13.0
evaluate>=0.4.0
wandb>=0.18.0

# Constrained decoding
outlines>=0.1.0

# Retrieval baseline
sentence-transformers>=3.0.0
rank-bm25>=0.2.2
